+++ modules/fitsfilereader/src/fitsfilereader.cpp
@@ -325,7 +325,7 @@ std::vector<float> FitsFileReader::readFitsFile(std::string filePath, int& nValu
         values[idx++] = tycho_v_err[i % nStars];
 
         // Read extra columns, if any. This will slow down the sorting tremendously!
-        for (size_t col = defaultCols; col < nColumnsRead; ++col) {
+        for (int col = defaultCols; col < nColumnsRead; ++col) {
             std::vector<float> vecData = std::move(tableContent[allColumnNames[col]]);
             values[idx++] = vecData[i];
         }
